django-app-graphql/requirements.txt
aniso8601==8.1.1
arrow==1.1.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
Django==3.2.4
django-appconf==1.0.4
django-currentuser==0.5.3
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-koldar-utils==2.8.1
django-polymorphic==3.0.0
djangorestframework==3.12.4
graphene==3.0b7
graphene-django==3.0.0b7
graphene-django-extras==0.5.1
graphql-core==3.1.5
graphql-relay==3.1.0
idna==2.10
inflect==5.3.0
inflection==0.5.1
jsonpath-ng==1.5.2
networkx==2.5.1
ply==3.11
promise==2.3
pydot==1.4.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
stringcase==1.2.0
text-unidecode==1.3
urllib3==1.26.5
